culture tweaks (#9)
switches the hegemony to unathi faction and the various regions to culture. adds faction options for the Ssen-Uuma and Rah'Zakeh. Adds religion options for the Markesheli and Aga-Eakhe. Removes all mentions of the SCG, replaced with either the Axis or Confederation. Removes all mentions of the ICCG and replaces them with lore on the history of the Gilgamesh Traverse.