Yet more psionics tweaks and balance part III
Fixes 'Arc' so that it should now actually function as a full welder. Adjusts the damage type from brute to burn, and increases the damage to be half of a lit welder's damage per strike.

Summoning 'Arc' now requires you to be targeting the hands while on harm intent.
Updated the descriptions of 'Arc' and 'Psiblade' to reflect the above change.

'Spark' now allows you to snap your fingers and light candles, incense cones and matches, as well as cigars and cigarettes.

Removes examine text for manifested psionic items.

Telekinesis can now be used to interact with light switches and airlock cycle buttons.

Adjusts the following psi-powers:
Energistics:
- Spark: cooldown 10 -> 20.
- Arc: maintain cost 3 -> 2.
- Flare: cooldown 65 - 60.
- Zorch: cooldown 65 -> 60.

Psychokinesis:
- Telekinesis: cooldown 25 -> 20.